# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Server Monitor API
  description: Monitor your server status
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /monitors:
    x-swagger-router-controller: monitors
    get:
      operationId: getMonitorList
      summary: Monitor Types
      description: |
        The Monitors endpoint returns information about the monitors available
        for a given server. The response includes the display name and other
        details about each monitor.
      tags:
        - Monitor
      responses:
        '200':
          description: An array of monitors
          schema:
            type: array
            items:
              $ref: "#/definitions/Monitor"
        '406':
          description: Negotiation not acceptable.
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
  "/monitors/{monitorId}":
    x-swagger-router-controller: monitors
    parameters:
      - name: monitorId
        type: string
        in: path
        description: ID of the monitor.
        required: true
    get:
      operationId: getMonitor
      summary: Information about a given monitor.
      description: Shows all the details regarding the monitor.
      tags:
        - Monitor
      responses:
        200:
          description: Details of the given monitor.
          schema:
            $ref: "#/definitions/Monitor"
        406:
          description: Negotiation not acceptable.
          schema:
            $ref: "#/definitions/Error"
        404:
          description: Monitor not found.
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Unexpected error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Metric:
    type: object
    properties:
      name:
        type: string
        description: Name of the metric.
      interval:
        type: integer
        format: int64
        description: Number of seconds between metric updates.
      retention:
        type: integer
        format: int64
        description: Number of seconds of TTL.
  Monitor:
    type: object
    properties:
      name:
        type: string
        description: Name of the monitor for a given server. For example, CPU or memory.
      description:
        type: string
        description: Description of monitor.
      metrics:
        type: array
        items:
          $ref: "#/definitions/Metric"
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string